feat: update contact form with new schedule and add phone field for improved user experience
