+++ i18n/locales/en.json
@@ -454,7 +454,7 @@
       "title": "Contact information",
       "email": "Email",
       "schedule": "Support hours",
-      "schedule_text": "24/7 - Support always available",
+      "schedule_text": "Monday to Saturday from 5:00 PM to 9:00 PM",
       "response": "Average response",
       "response_text": "Less than 2 hours"
     },
