+++ i18n/locales/es.json
@@ -453,7 +453,7 @@
       "title": "Información de contacto",
       "email": "Email",
       "schedule": "Horario de atención",
-      "schedule_text": "24/7 - Soporte disponible siempre",
+      "schedule_text": "Lunes a Sábado de 17:00 a 21:00",
       "response": "Respuesta promedio",
       "response_text": "Menos de 2 horas"
     },
